set/get2ByteInt gefixt

+++ KnxTpUart/KnxTelegram.h
@@ -87,10 +87,10 @@ class KnxTelegram {
 		bool getBool();
 		void set2ByteFloatValue(float value);
 		float get2ByteFloatValue();
-		void set2ByteIntValue(float value);
+		void set2ByteIntValue(int value);
         int get1ByteIntValue();
 		void set1ByteIntValue(int value);
-		float get2ByteIntValue();
+		int get2ByteIntValue();
         void set4ByteFloatValue(float value);
         float get4ByteFloatValue();
         void set14ByteValue(String value);
@@ -115,4 +115,3 @@ class KnxTelegram {
 };
 
 #endif
-
